+++ HackrNewsTests/Live Hackr News API/Helpers/HTTPClientSpy.swift
@@ -1,0 +1,41 @@
+//
+//  Copyright © 2020 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Foundation
+import HackrNews
+
+final class HTTPClientSpy: HTTPClient {
+    private struct Task: HTTPClientTask {
+        let completion: () -> Void
+
+        func cancel() {
+            completion()
+        }
+    }
+
+    var messages = [(url: URL, completion: (HTTPClient.Result) -> Void)]()
+    var requestedURLs: [URL] { messages.map(\.url) }
+    var cancelledURLs = [URL]()
+
+    func get(from url: URL, completion: @escaping (HTTPClient.Result) -> Void) -> HTTPClientTask {
+        messages.append((url, completion))
+        return Task { [weak self] in
+            self?.cancelledURLs.append(url)
+        }
+    }
+
+    func complete(with error: Error, at index: Int = 0) {
+        messages[index].completion(.failure(error))
+    }
+
+    func complete(with code: Int, data: Data, at index: Int = 0) {
+        let response = HTTPURLResponse(
+            url: requestedURLs[index],
+            statusCode: code,
+            httpVersion: nil,
+            headerFields: nil
+        )!
+        messages[index].completion(.success((data, response)))
+    }
+}
